Fix a bunch of type mismatch-related warnings (Jan Kiszka).
Fix a typo in my previous comming (spotted by Laurent Desnouges).


git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@4877 c046a42c-6fe2-441c-8c8c-71466251a162

+++ linux-user/flatload.c
@@ -349,9 +349,9 @@ void old_reloc(struct lib_info *libinfo, uint32_t rl)
         reloc_type = rl >> 30;
         /* ??? How to handle this?  */
 #if defined(CONFIG_COLDFIRE)
-	ptr = (uint32_t *) (libinfo->start_code + offset);
+	ptr = (uint32_t *) ((unsigned long) libinfo->start_code + offset);
 #else
-	ptr = (uint32_t *) (libinfo->start_data + offset);
+	ptr = (uint32_t *) ((unsigned long) libinfo->start_data + offset);
 #endif
 
 #ifdef DEBUG
@@ -670,7 +670,7 @@ static int load_flat_file(struct linux_binprm * bprm,
     }
 
     /* zero the BSS.  */
-    memset((void*)(datapos + data_len), 0, bss_len);
+    memset((void *)((unsigned long)datapos + data_len), 0, bss_len);
 
     return 0;
 }
